Split NewFileMetadata from pkg/chunked/compressor
In addition to the existing use when creating a TOC from tar data,
we will also need it when parsing TOC and tar-split data.

Should not change behavior.
